+++ Assets/Scripts/Sounds/MusicSystem.cs
@@ -8,45 +8,63 @@ using Random = UnityEngine.Random;
 public class MusicSystem : MonoBehaviour
 {
     [SerializeField] private List<Music> _musics;
+    public static ActiveMusic MainMusic { get; private set; }
 
-    public static ActiveMusic MainSoundtrack { get; private set; }
+    public static MusicSystem Singleton { get; private set; }
 
-    public static MusicSystem Singleton()
+    private void OnDestroy()
     {
-        return FindFirstObjectByType<MusicSystem>(FindObjectsInactive.Include);
+        Singleton = null;
+        MainMusic = null;
+    }
+
+    private void Awake()
+    {
+        Singleton = this;
     }
 
     public static int GetRandomMusicIndex()
     {
-        MusicSystem current = Singleton();
-
-        var sceneMusics = current._musics.FindAll(music => music.scene == SceneManager.GetActiveScene().buildIndex);
+        var sceneMusics = Singleton._musics.FindAll(music => music.scene == SceneManager.GetActiveScene().buildIndex);
         if (sceneMusics.Count == 0)
         {
             Debug.LogWarning("Cant find musics for this scene");
             return -1;
         }
 
-        return current._musics.IndexOf(sceneMusics[Random.Range(0, sceneMusics.Count)]);
+        return Singleton._musics.IndexOf(sceneMusics[Random.Range(0, sceneMusics.Count)]);
     }
 
-    public static ActiveMusic StartMusic(MusicGameState state, int index, float offset = 0f)
+    public static void StartMusic(int index, float offset = 0f)
     {
-        if (index == -1) return null;
-        MusicSystem current = Singleton();
+        if (index == -1)
+        {
+            MainMusic = null;
+            return;
+        }
 
-        AudioClip target = current._musics[index].music;
+        AudioClip target = Singleton._musics[index].music;
         AudioSource source = SoundSystem.PlaySound(new SoundTransporter(target), new SoundPositioner(Vector3.zero), SoundType.Music, volume: 0.525f, enableFade: false);
         source.time = offset;
 
-        ActiveMusic newInstance = new ActiveMusic()
+        ActiveMusic newInstance = new()
         {
             Current = target,
             Source = source
         };
 
-        MainSoundtrack = newInstance;
-        return newInstance;
+        MainMusic = newInstance;
+    }
+
+    private void Update()
+    {
+        if (MainMusic != null && MainMusic.Source && GameInfo.Singleton.CurrentMusicOffset > 0)
+        {
+            var speed = 1f + (GameInfo.Singleton.CurrentMusicOffset - MainMusic.Source.time);
+
+            if (speed <= 0.25f || speed >= 1.75f) MainMusic.Source.time = GameInfo.Singleton.CurrentMusicOffset;
+            else MainMusic.Source.pitch = speed;
+        }
     }
 }